+++ user-service/src/test/java/kr/mybrary/userservice/authentication/domain/oauth2/service/AppleOAuth2UserServiceTest.java
@@ -37,6 +37,8 @@ class AppleOAuth2UserServiceTest {
     @Mock
     AppleOAuth2ServiceClient appleOAuth2ServiceClient;
     @Mock
+    AppleOAuth2UtilService appleOAuth2UtilService;
+    @Mock
     UserRepository userRepository;
     @Mock
     PasswordEncoder passwordEncoder;
@@ -55,11 +57,7 @@ class AppleOAuth2UserServiceTest {
     @DisplayName("이미 가입된 애플 소셜 로그인 회원을 로그인한다.")
     void authenticateRegisteredUser() {
         // given
-        appleOAuth2UserService.setAPPLE_CLIENT_SECRET("authKeyFile/keyId/teamId");
-        appleOAuth2UserService.setAPPLE_CLIENT_ID("clientId");
-        appleOAuth2UserService.setAPPLE_AUTHORIZATION_GRANT_TYPE("authorization_code");
-        appleOAuth2UserService.setAPPLE_REDIRECT_URI("https://user.mybrary.kr/oauth2/apple/callback");
-
+        setUpAppleInfo();
         MockHttpServletRequest request = new MockHttpServletRequest();
         request.setParameter("code", "code");
 
@@ -71,6 +69,7 @@ class AppleOAuth2UserServiceTest {
                         .refresh_token("refresh_token")
                         .id_token(ID_TOKEN_SAMPLE)
                         .build());
+        given(appleOAuth2UtilService.createAppleClientSecret(anyString(), anyString())).willReturn("clientSecret");
         given(objectMapper.convertValue(any(), (Class<Object>) any())).willReturn(mock(JSONObject.class));
         given(userRepository.findBySocialTypeAndSocialId(any(), any())).willReturn(Optional.ofNullable(UserFixture.APPLE_USER.getUser()));
         given(jwtUtil.createAccessToken(any(), any())).willReturn("accessToken");
@@ -83,10 +82,18 @@ class AppleOAuth2UserServiceTest {
         // then
         assertEquals("kr.mybrary://?Authorization=accessToken&Authorization-Refresh=refreshToken", redirectUrl);
         verify(appleOAuth2ServiceClient, times(1)).getAppleToken(anyString(), anyString(), anyString(), anyString(), anyString());
+        verify(appleOAuth2UtilService, times(1)).createAppleClientSecret(anyString(), anyString());
         verify(userRepository, times(1)).findBySocialTypeAndSocialId(any(), any());
         verify(jwtUtil, times(1)).createAccessToken(any(), any());
         verify(jwtUtil, times(1)).createRefreshToken(any());
         verify(redisUtil, times(1)).set(any(), any(), any());
+    }
+
+    private void setUpAppleInfo() {
+        appleOAuth2UserService.setAPPLE_CLIENT_ID("clientId");
+        appleOAuth2UserService.setAPPLE_CLIENT_SECRET("clientSecret");
+        appleOAuth2UserService.setAPPLE_AUTHORIZATION_GRANT_TYPE("authorization_code");
+        appleOAuth2UserService.setAPPLE_REDIRECT_URI("https://user.mybrary.kr/oauth2/apple/callback");
     }
 
     @Test
@@ -110,11 +117,7 @@ class AppleOAuth2UserServiceTest {
     @DisplayName("이미 가입된 애플 소셜 로그인 회원을 로그인할 때 사용자를 찾을 수 없으면 예외가 발생한다.")
     void authenticateRegisteredUserNotFound() {
         // given
-        appleOAuth2UserService.setAPPLE_CLIENT_SECRET("authKeyFile/keyId/teamId");
-        appleOAuth2UserService.setAPPLE_CLIENT_ID("clientId");
-        appleOAuth2UserService.setAPPLE_AUTHORIZATION_GRANT_TYPE("authorization_code");
-        appleOAuth2UserService.setAPPLE_REDIRECT_URI("https://user.mybrary.kr/oauth2/apple/callback");
-
+        setUpAppleInfo();
         MockHttpServletRequest request = new MockHttpServletRequest();
         request.setParameter("code", "code");
 
@@ -126,6 +129,7 @@ class AppleOAuth2UserServiceTest {
                         .refresh_token("refresh_token")
                         .id_token(ID_TOKEN_SAMPLE)
                         .build());
+        given(appleOAuth2UtilService.createAppleClientSecret(anyString(), anyString())).willReturn("clientSecret");
         given(objectMapper.convertValue(any(), (Class<Object>) any())).willReturn(mock(JSONObject.class));
         given(userRepository.findBySocialTypeAndSocialId(any(), any())).willReturn(Optional.empty());
 
@@ -135,17 +139,18 @@ class AppleOAuth2UserServiceTest {
                 .hasFieldOrPropertyWithValue("status", 404)
                 .hasFieldOrPropertyWithValue("errorCode", "A-05")
                 .hasFieldOrPropertyWithValue("errorMessage", "Apple 사용자를 찾을 수 없습니다.");
+
+        verify(appleOAuth2ServiceClient, times(1)).getAppleToken(anyString(), anyString(), anyString(), anyString(), anyString());
+        verify(appleOAuth2UtilService, times(1)).createAppleClientSecret(anyString(), anyString());
+        verify(objectMapper, times(1)).convertValue(any(), (Class<Object>) any());
+        verify(userRepository, times(1)).findBySocialTypeAndSocialId(any(), any());
     }
 
     @Test
     @DisplayName("최초 애플 소셜 로그인 회원을 회원 가입 후 로그인한다.")
     void authenticateNewUser() throws Exception {
         // given
-        appleOAuth2UserService.setAPPLE_CLIENT_SECRET("authKeyFile/keyId/teamId");
-        appleOAuth2UserService.setAPPLE_CLIENT_ID("clientId");
-        appleOAuth2UserService.setAPPLE_AUTHORIZATION_GRANT_TYPE("authorization_code");
-        appleOAuth2UserService.setAPPLE_REDIRECT_URI("https://user.mybrary.kr/oauth2/apple/callback");
-
+        setUpAppleInfo();
         MockHttpServletRequest request = new MockHttpServletRequest();
         request.setParameter("code", "code");
         request.setParameter("user", "user");
@@ -163,6 +168,7 @@ class AppleOAuth2UserServiceTest {
                         .refresh_token("refresh_token")
                         .id_token(ID_TOKEN_SAMPLE)
                         .build());
+        given(appleOAuth2UtilService.createAppleClientSecret(anyString(), anyString())).willReturn("clientSecret");
         given(passwordEncoder.encode(any())).willReturn("encodedPassword");
         given(userRepository.save(any())).willReturn(UserFixture.APPLE_USER.getUser());
         given(jwtUtil.createAccessToken(any(), any())).willReturn("accessToken");
@@ -177,6 +183,7 @@ class AppleOAuth2UserServiceTest {
         verify(objectMapper, times(1)).readValue(eq("user"), eq(AppleOAuth2UserInfo.class));
         verify(objectMapper, times(1)).convertValue(any(), (Class<Object>) any());
         verify(appleOAuth2ServiceClient, times(1)).getAppleToken(anyString(), anyString(), anyString(), anyString(), anyString());
+        verify(appleOAuth2UtilService, times(1)).createAppleClientSecret(anyString(), anyString());
         verify(passwordEncoder, times(1)).encode(any());
         verify(userRepository, times(1)).save(any());
         verify(jwtUtil, times(1)).createAccessToken(any(), any());
@@ -227,13 +234,10 @@ class AppleOAuth2UserServiceTest {
     @DisplayName("애플 소셜 로그인 회원의 토큰을 만료하고 연동을 해제한다.")
     void withdrawApple() {
         // given
-        appleOAuth2UserService.setAPPLE_CLIENT_SECRET("authKeyFile/keyId/teamId");
-        appleOAuth2UserService.setAPPLE_CLIENT_ID("clientId");
-        appleOAuth2UserService.setAPPLE_AUTHORIZATION_GRANT_TYPE("authorization_code");
-        appleOAuth2UserService.setAPPLE_REDIRECT_URI("https://user.mybrary.kr/oauth2/apple/callback");
-
+        setUpAppleInfo();
         given(redisUtil.get(any())).willReturn("refreshToken");
         doNothing().when(appleOAuth2ServiceClient).revokeAppleToken(any(), any(), any(), any());
+        given(appleOAuth2UtilService.createAppleClientSecret(anyString(), anyString())).willReturn("clientSecret");
         doNothing().when(redisUtil).delete(any());
 
         // when
@@ -242,6 +246,7 @@ class AppleOAuth2UserServiceTest {
         // then
         verify(redisUtil, times(1)).get(any());
         verify(appleOAuth2ServiceClient, times(1)).revokeAppleToken(any(), any(), any(), any());
+        verify(appleOAuth2UtilService, times(1)).createAppleClientSecret(anyString(), anyString());
         verify(redisUtil, times(1)).delete(any());
     }
